Fix a crash in HungRendererView. |tab_contents_| can be set to NULL when TabContents is destroyed, when RenderViewIsGone and when the tab becomes responsive again. If we happen to press the button just after this happens, we'll get a crash.
BUG=http://crbug.com/52293
TEST=None

Review URL: http://codereview.chromium.org/3165022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56419 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2627b1ed1f32f89609797290d6222b748b657757

+++ chrome/browser/views/hung_renderer_view.cc
@@ -318,9 +318,11 @@ views::View* HungRendererDialogView::GetContentsView() {
 void HungRendererDialogView::ButtonPressed(
     views::Button* sender, const views::Event& event) {
   if (sender == kill_button_) {
-    // Kill the process.
-    TerminateProcess(contents_->GetRenderProcessHost()->GetHandle(),
-                     ResultCodes::HUNG);
+    if (!contents_) {
+      // Kill the process.
+      TerminateProcess(contents_->GetRenderProcessHost()->GetHandle(),
+                       ResultCodes::HUNG);
+    }
   }
 }
